Introduce AspNetCoreApiDescriptionModelProviderOptions.
Resolve #5921
