+++ framework/src/Volo.Abp.AspNetCore.Mvc/Volo/Abp/AspNetCore/Mvc/AspNetCoreApiDescriptionModelProvider.cs
@@ -25,17 +25,20 @@ namespace Volo.Abp.AspNetCore.Mvc
     {
         public ILogger<AspNetCoreApiDescriptionModelProvider> Logger { get; set; }
 
+        private readonly AspNetCoreApiDescriptionModelProviderOptions _options;
         private readonly IApiDescriptionGroupCollectionProvider _descriptionProvider;
-        private readonly AbpAspNetCoreMvcOptions _options;
+        private readonly AbpAspNetCoreMvcOptions _abpAspNetCoreMvcOptions;
         private readonly AbpApiDescriptionModelOptions _modelOptions;
 
         public AspNetCoreApiDescriptionModelProvider(
+            IOptions<AspNetCoreApiDescriptionModelProviderOptions> options,
             IApiDescriptionGroupCollectionProvider descriptionProvider,
-            IOptions<AbpAspNetCoreMvcOptions> options,
+            IOptions<AbpAspNetCoreMvcOptions> abpAspNetCoreMvcOptions,
             IOptions<AbpApiDescriptionModelOptions> modelOptions)
         {
-            _descriptionProvider = descriptionProvider;
             _options = options.Value;
+            _descriptionProvider = descriptionProvider;
+            _abpAspNetCoreMvcOptions = abpAspNetCoreMvcOptions.Value;
             _modelOptions = modelOptions.Value;
 
             Logger = NullLogger<AspNetCoreApiDescriptionModelProvider>.Instance;
@@ -82,14 +85,14 @@ namespace Volo.Abp.AspNetCore.Mvc
 
             var controllerModel = moduleModel.GetOrAddController(
                 controllerType.FullName,
-                CalculateControllerName(controllerType, setting),
+                _options.ControllerNameGenerator(controllerType, setting),
                 controllerType,
                 _modelOptions.IgnoredInterfaces
             );
 
             var method = apiDescription.ActionDescriptor.GetMethodInfo();
 
-            var uniqueMethodName = GetUniqueActionName(method);
+            var uniqueMethodName = _options.ActionNameGenerator(method);
             if (controllerModel.Actions.ContainsKey(uniqueMethodName))
             {
                 Logger.LogWarning(
@@ -117,44 +120,6 @@ namespace Volo.Abp.AspNetCore.Mvc
             }
 
             AddParameterDescriptionsToModel(actionModel, method, apiDescription);
-        }
-
-        private static string CalculateControllerName(Type controllerType, ConventionalControllerSetting setting)
-        {
-            var controllerName = controllerType.Name.RemovePostFix("Controller")
-                .RemovePostFix(ApplicationService.CommonPostfixes);
-
-            if (setting?.UrlControllerNameNormalizer != null)
-            {
-                controllerName =
-                    setting.UrlControllerNameNormalizer(
-                        new UrlControllerNameNormalizerContext(setting.RootPath, controllerName));
-            }
-
-            return controllerName;
-        }
-
-        private static string GetUniqueActionName(MethodInfo method)
-        {
-            var methodNameBuilder = new StringBuilder(method.Name);
-
-            var parameters = method.GetParameters();
-            if (parameters.Any())
-            {
-                methodNameBuilder.Append("By");
-
-                for (var i = 0; i < parameters.Length; i++)
-                {
-                    if (i > 0)
-                    {
-                        methodNameBuilder.Append("And");
-                    }
-
-                    methodNameBuilder.Append(parameters[i].Name.ToPascalCase());
-                }
-            }
-
-            return methodNameBuilder.ToString();
         }
 
         private static List<string> GetSupportedVersions(Type controllerType, MethodInfo method,
@@ -377,7 +342,7 @@ namespace Volo.Abp.AspNetCore.Mvc
         [CanBeNull]
         private ConventionalControllerSetting FindSetting(Type controllerType)
         {
-            foreach (var controllerSetting in _options.ConventionalControllers.ConventionalControllerSettings)
+            foreach (var controllerSetting in _abpAspNetCoreMvcOptions.ConventionalControllers.ConventionalControllerSettings)
             {
                 if (controllerSetting.ControllerTypes.Contains(controllerType))
                 {
